modify attendance form

+++ service/src/main/java/com/nicico/training/iservice/ITclassService.java
@@ -44,7 +44,7 @@ public interface ITclassService {
     void addStudents(StudentDTO.Delete request, Long classId);
 
     @Transactional(readOnly = true)
-    Float sessionsHourSum(Long classId);
+    Long sessionsHourSum(Long classId);
 
     @Transactional
     Long getEndGroup(Long courseId, Long termId);
